Move ChainDataIndex to its own utils module
