+++ crates/storage/store/engines.rs
@@ -3,3 +3,4 @@ pub mod in_memory;
 #[cfg(feature = "libmdbx")]
 pub mod libmdbx;
 pub mod redb;
+mod utils;
