+++ crates/storage/store/engines/libmdbx.rs
@@ -1,4 +1,5 @@
 use super::api::StoreEngine;
+use super::utils::ChainDataIndex;
 use crate::error::StoreError;
 use crate::rlp::{
     AccountCodeHashRLP, AccountCodeRLP, BlockBodyRLP, BlockHashRLP, BlockHeaderRLP, BlockRLP,
@@ -20,7 +21,6 @@ use libmdbx::{
     table_info,
 };
 use libmdbx::{DatabaseOptions, Mode, ReadWriteOptions};
-use redb::TypeName;
 use serde_json;
 use std::fmt::{Debug, Formatter};
 use std::path::Path;
@@ -560,87 +560,11 @@ impl From<AccountStorageValueBytes> for U256 {
     }
 }
 
-/// Represents the key for each unique value of the chain data stored in the db
-// (TODO: Remove this comment once full) Will store chain-specific data such as chain id and latest finalized/pending/safe block number
-#[derive(Debug, Copy, Clone)]
-pub enum ChainDataIndex {
-    ChainConfig = 0,
-    EarliestBlockNumber = 1,
-    FinalizedBlockNumber = 2,
-    SafeBlockNumber = 3,
-    LatestBlockNumber = 4,
-    PendingBlockNumber = 5,
-    // TODO (#307): Remove TotalDifficulty.
-    LatestTotalDifficulty = 6,
-}
-
-impl From<u8> for ChainDataIndex {
-    fn from(value: u8) -> Self {
-        match value {
-            x if x == ChainDataIndex::ChainConfig as u8 => ChainDataIndex::ChainConfig,
-            x if x == ChainDataIndex::EarliestBlockNumber as u8 => {
-                ChainDataIndex::EarliestBlockNumber
-            }
-            x if x == ChainDataIndex::FinalizedBlockNumber as u8 => {
-                ChainDataIndex::FinalizedBlockNumber
-            }
-            x if x == ChainDataIndex::SafeBlockNumber as u8 => ChainDataIndex::SafeBlockNumber,
-            x if x == ChainDataIndex::LatestBlockNumber as u8 => ChainDataIndex::LatestBlockNumber,
-            x if x == ChainDataIndex::PendingBlockNumber as u8 => {
-                ChainDataIndex::PendingBlockNumber
-            }
-            x if x == ChainDataIndex::LatestTotalDifficulty as u8 => {
-                ChainDataIndex::LatestTotalDifficulty
-            }
-            _ => panic!("Invalid value when casting to ChainDataIndex: {}", value),
-        }
-    }
-}
-
 impl Encodable for ChainDataIndex {
     type Encoded = [u8; 4];
 
     fn encode(self) -> Self::Encoded {
         (self as u32).encode()
-    }
-}
-
-impl redb::Value for ChainDataIndex {
-    type SelfType<'a> = ChainDataIndex
-    where
-        Self: 'a;
-
-    type AsBytes<'a> = [u8; 1]
-    where
-        Self: 'a;
-
-    fn fixed_width() -> Option<usize> {
-        None
-    }
-
-    fn from_bytes<'a>(data: &'a [u8]) -> Self::SelfType<'a>
-    where
-        Self: 'a,
-    {
-        data[0].into()
-    }
-
-    fn as_bytes<'a, 'b: 'a>(value: &'a Self::SelfType<'b>) -> Self::AsBytes<'a>
-    where
-        Self: 'a,
-        Self: 'b,
-    {
-        [*value as u8]
-    }
-
-    fn type_name() -> redb::TypeName {
-        TypeName::new("ChainDataIndex")
-    }
-}
-
-impl redb::Key for ChainDataIndex {
-    fn compare(data1: &[u8], data2: &[u8]) -> std::cmp::Ordering {
-        data1.cmp(data2)
     }
 }
